rename package name libbun.parser.ssa2 => libbun.parser.ssa
