+++ java/src/libbun/encode/SSACGenerator.java
@@ -47,8 +47,8 @@ import libbun.parser.ast.ZSetterNode;
 import libbun.parser.ast.ZThrowNode;
 import libbun.parser.ast.ZTryNode;
 import libbun.parser.ast.ZVarBlockNode;
-import libbun.parser.ssa2.PHINode;
-import libbun.parser.ssa2.SSAConverter;
+import libbun.parser.ssa.PHINode;
+import libbun.parser.ssa.SSAConverter;
 import libbun.type.ZClassField;
 import libbun.type.ZClassType;
 import libbun.type.ZFunc;
